Move dataset builder code to top level directory
