+++ src/dataset-builder/Breadcrumbs.test.ts
@@ -1,7 +1,8 @@
 import { screen } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
-import { DatasetBuilderBreadcrumbs } from 'src/pages/library/datasetBuilder/Breadcrumbs';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
+
+import { DatasetBuilderBreadcrumbs } from './Breadcrumbs';
 
 describe('Breadcrumbs', () => {
   it('renders a single breadcrumb', () => {
