+++ src/dataset-builder/ConceptSelector.test.ts
@@ -2,8 +2,9 @@ import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { h } from 'react-hyperscript-helpers';
 import { DatasetBuilder, DatasetBuilderContract, getConceptForId } from 'src/libs/ajax/DatasetBuilder';
-import { ConceptSelector } from 'src/pages/library/datasetBuilder/ConceptSelector';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
+
+import { ConceptSelector } from './ConceptSelector';
 
 type DatasetBuilderExports = typeof import('src/libs/ajax/DatasetBuilder');
 jest.mock('src/libs/ajax/DatasetBuilder', (): DatasetBuilderExports => {
