+++ src/dataset-builder/ConceptSelector.ts
@@ -8,7 +8,8 @@ import { icon } from 'src/components/icons';
 import { TreeGrid } from 'src/components/TreeGrid';
 import { SnapshotBuilderConcept as Concept } from 'src/libs/ajax/DataRepo';
 import { DatasetBuilder, getConceptForId } from 'src/libs/ajax/DatasetBuilder';
-import { PAGE_PADDING_HEIGHT, PAGE_PADDING_WIDTH } from 'src/pages/library/datasetBuilder/constants';
+
+import { PAGE_PADDING_HEIGHT, PAGE_PADDING_WIDTH } from './constants';
 
 const getChildren = async (concept: Concept): Promise<Concept[]> => {
   const result = await DatasetBuilder().getConcepts(concept);
