+++ src/dataset-builder/DatasetBuilderHeader.ts
@@ -3,8 +3,9 @@ import { Link } from 'src/components/common';
 import { DatasetModel } from 'src/libs/ajax/DataRepo';
 import colors from 'src/libs/colors';
 import * as Nav from 'src/libs/nav';
-import { DatasetBuilderBreadcrumbs } from 'src/pages/library/datasetBuilder/Breadcrumbs';
-import { PAGE_PADDING_HEIGHT, PAGE_PADDING_WIDTH } from 'src/pages/library/datasetBuilder/constants';
+
+import { DatasetBuilderBreadcrumbs } from './Breadcrumbs';
+import { PAGE_PADDING_HEIGHT, PAGE_PADDING_WIDTH } from './constants';
 
 type DatasetBuilderHeaderProps = { datasetDetails: DatasetModel };
