+++ src/libs/routes.js
@@ -17,8 +17,8 @@ import * as ImportWorkflow from 'src/pages/ImportWorkflow/ImportWorkflow';
 import * as LandingPage from 'src/pages/LandingPage';
 import * as Code from 'src/pages/library/Code';
 import * as CreateDataset from 'src/pages/library/data-catalog/create-dataset/CreateDataset';
-import * as DatasetBuilder from 'src/pages/library/datasetBuilder/DatasetBuilder';
-import * as DatasetBuilderDetails from 'src/pages/library/datasetBuilder/DatasetBuilderDetails';
+import * as DatasetBuilder from 'src/pages/library/dataset-builder/DatasetBuilder';
+import * as DatasetBuilderDetails from 'src/pages/library/dataset-builder/DatasetBuilderDetails';
 import * as Datasets from 'src/pages/library/Datasets';
 import * as DataExplorer from 'src/pages/library/datasets/DataExplorer';
 import * as Showcase from 'src/pages/library/Showcase';
